+++ src/Form/Element/Text.php
@@ -4,6 +4,22 @@ namespace SleepingOwl\Admin\Form\Element;
 
 class Text extends NamedFormElement
 {
+
+    /**
+     * @var string
+     */
+    protected $view = 'form.element.text';
+
+    /**
+     * @var bool
+     */
+    protected $dataList = false;
+
+    /**
+     * @var array
+     */
+    protected $options = [];
+
     public function __construct($path, $label = null)
     {
         parent::__construct($path, $label);
@@ -14,8 +30,40 @@ class Text extends NamedFormElement
         ]);
     }
 
+
     /**
-     * @var string
+     * @param array
+     * @return $this
      */
-    protected $view = 'form.element.text';
+    public function setOptions(array $options): Text
+    {
+        $this->options = $options;
+        $this->dataList = true;
+
+        return $this;
+    }
+
+
+    /**
+     * @return array
+     */
+    public function toArray(): array
+    {
+        if ($this->dataList) {
+            $this->setHtmlAttributes([
+                'list' => $this->getId().'Datalist',
+            ]);
+
+            return
+                parent::toArray() + [
+                    'datalistOptions' => $this->options
+                ];
+        }
+
+        return parent::toArray();
+    }
+
+
+
+
 }